+++ example/lib/main.dart
@@ -1,7 +1,9 @@
+// Usually, I would prefer splitting up an app into smaller files, but as this
+// is an example app for a published plugin, it's better to have everything in
+// one file so that all of the examples are visible on https://pub.dev/packages/esptouch_flutter/example
+
 import 'dart:async';
 
-import 'package:esptouch_example/task_parameter_details.dart';
-import 'package:esptouch_example/wifi_info.dart';
 import 'package:esptouch_flutter/esptouch_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -13,52 +15,41 @@ class MyApp extends StatefulWidget {
   _MyAppState createState() => _MyAppState();
 }
 
-const helperSSID =
-    "SSID is the technical term for a network name. When you set up a wireless home network, you give it a name to distinguish it from other networks in your neighbourhood.";
-const helperBSSID =
-    "BSSID is the MAC address of the wireless access point (router).";
-const helperPassword = "The password of the Wi-Fi network";
-
 class _MyAppState extends State<MyApp> {
-  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  final TextEditingController _ssid = TextEditingController();
-  final TextEditingController _bssid = TextEditingController();
-  final TextEditingController _password = TextEditingController();
-  final TextEditingController _expectedTaskResults =
-      TextEditingController(); // TODO; is it the same as threshold?
-  final TextEditingController _intervalGuideCode = TextEditingController();
-  final TextEditingController _intervalDataCode = TextEditingController();
-  final TextEditingController _timeoutGuideCode = TextEditingController();
-  final TextEditingController _timeoutDataCode = TextEditingController();
-  final TextEditingController _repeat = TextEditingController();
-
-  // final TextEditingController _oneLength = TextEditingController();
-  // final TextEditingController _macLength = TextEditingController();
-  // final TextEditingController _ipLength = TextEditingController();
-  final TextEditingController _portListening = TextEditingController();
-  final TextEditingController _portTarget = TextEditingController();
-  final TextEditingController _waitUdpReceiving = TextEditingController();
-  final TextEditingController _waitUdpSending = TextEditingController();
-  final TextEditingController _thresholdSucBroadcastCount =
+  final GlobalKey<FormState> formKey = GlobalKey<FormState>();
+  final TextEditingController ssid = TextEditingController();
+  final TextEditingController bssid = TextEditingController();
+  final TextEditingController password = TextEditingController();
+  final TextEditingController expectedTaskResults = TextEditingController();
+  final TextEditingController intervalGuideCode = TextEditingController();
+  final TextEditingController intervalDataCode = TextEditingController();
+  final TextEditingController timeoutGuideCode = TextEditingController();
+  final TextEditingController timeoutDataCode = TextEditingController();
+  final TextEditingController repeat = TextEditingController();
+  final TextEditingController portListening = TextEditingController();
+  final TextEditingController portTarget = TextEditingController();
+  final TextEditingController waitUdpReceiving = TextEditingController();
+  final TextEditingController waitUdpSending = TextEditingController();
+  final TextEditingController thresholdSucBroadcastCount =
       TextEditingController();
-  ESPTouchPacket _packet = ESPTouchPacket.broadcast;
+  ESPTouchPacket packet = ESPTouchPacket.broadcast;
 
   @override
   void dispose() {
-    _ssid.dispose();
-    _bssid.dispose();
-    _password.dispose();
-    _expectedTaskResults.dispose();
-    _intervalGuideCode.dispose();
-    _intervalDataCode.dispose();
-    _timeoutGuideCode.dispose();
-    _timeoutDataCode.dispose();
-    _repeat.dispose();
-    _portListening.dispose();
-    _portTarget.dispose();
-    _waitUdpReceiving.dispose();
-    _waitUdpSending.dispose();
-    _thresholdSucBroadcastCount.dispose();
+    ssid.dispose();
+    bssid.dispose();
+    password.dispose();
+    expectedTaskResults.dispose();
+    intervalGuideCode.dispose();
+    intervalDataCode.dispose();
+    timeoutGuideCode.dispose();
+    timeoutDataCode.dispose();
+    repeat.dispose();
+    portListening.dispose();
+    portTarget.dispose();
+    waitUdpReceiving.dispose();
+    waitUdpSending.dispose();
+    thresholdSucBroadcastCount.dispose();
     super.dispose();
   }
 
@@ -66,13 +57,6 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primaryColor: Colors.deepOrange,
-        buttonTheme: ButtonThemeData(
-          buttonColor: Colors.deepOrange,
-          textTheme: ButtonTextTheme.primary,
-        ),
-      ),
       home: Scaffold(
         appBar: AppBar(
           title: Text(
@@ -85,11 +69,7 @@ class _MyAppState extends State<MyApp> {
         ),
         // Using builder to get context without creating new widgets
         //  https://docs.flutter.io/flutter/material/Scaffold/of.html
-        body: Builder(builder: (BuildContext context) {
-          return Center(
-            child: form(context),
-          );
-        }),
+        body: Builder(builder: (context) => Center(child: form(context))),
       ),
     );
   }
@@ -97,96 +77,60 @@ class _MyAppState extends State<MyApp> {
   bool fetchingWifiInfo = false;
 
   void fetchWifiInfo() async {
-    setState(() {
-      fetchingWifiInfo = true;
-    });
+    setState(() => fetchingWifiInfo = true);
     try {
-      _ssid.text = await ssid;
-      _bssid.text = await bssid;
+      ssid.text = await SimpleWifiInfo.ssid ?? '';
+      bssid.text = await SimpleWifiInfo.bssid ?? '';
     } finally {
-      setState(() {
-        fetchingWifiInfo = false;
-      });
+      setState(() => fetchingWifiInfo = false);
     }
   }
 
   createTask() {
-    final taskParameter = ESPTouchTaskParameter();
-    if (_intervalGuideCode.text.isNotEmpty) {
-      taskParameter.intervalGuideCode =
-          Duration(milliseconds: int.parse(_intervalGuideCode.text));
-    }
-    if (_intervalDataCode.text.isNotEmpty) {
-      taskParameter.intervalDataCode =
-          Duration(milliseconds: int.parse(_intervalDataCode.text));
-    }
-    if (_timeoutGuideCode.text.isNotEmpty) {
-      taskParameter.timeoutGuideCode =
-          Duration(milliseconds: int.parse(_timeoutGuideCode.text));
-    }
-    if (_timeoutDataCode.text.isNotEmpty) {
-      taskParameter.timeoutDataCode =
-          Duration(milliseconds: int.parse(_timeoutDataCode.text));
-    }
-    if (_repeat.text.isNotEmpty) {
-      taskParameter.repeat = int.parse(_repeat.text);
-    }
-    if (_portListening.text.isNotEmpty) {
-      taskParameter.portListening = int.parse(_portListening.text);
-    }
-    if (_portTarget.text.isNotEmpty) {
-      taskParameter.portTarget = int.parse(_portTarget.text);
-    }
-    if (_waitUdpSending.text.isNotEmpty) {
-      taskParameter.waitUdpSending =
-          Duration(milliseconds: int.parse(_waitUdpSending.text));
-    }
-    if (_waitUdpReceiving.text.isNotEmpty) {
-      taskParameter.waitUdpReceiving =
-          Duration(milliseconds: int.parse(_waitUdpReceiving.text));
-    }
-    if (_thresholdSucBroadcastCount.text.isNotEmpty) {
-      taskParameter.thresholdSucBroadcastCount =
-          int.parse(_thresholdSucBroadcastCount.text);
-    }
-    if (_expectedTaskResults.text.isNotEmpty) {
-      taskParameter.expectedTaskResults = int.parse(_expectedTaskResults.text);
+    Duration? durationTryParse(String milliseconds) {
+      final parsed = int.tryParse(milliseconds);
+      return parsed != null ? Duration(milliseconds: parsed) : null;
     }
 
     return ESPTouchTask(
-      ssid: _ssid.text,
-      bssid: _bssid.text,
-      password: _password.text,
-      packet: _packet,
-      taskParameter: taskParameter,
+      ssid: ssid.text,
+      bssid: bssid.text,
+      password: password.text,
+      packet: packet,
+      taskParameter: ESPTouchTaskParameter().copyWith(
+        intervalGuideCode: durationTryParse(intervalGuideCode.text),
+        intervalDataCode: durationTryParse(intervalDataCode.text),
+        timeoutGuideCode: durationTryParse(timeoutGuideCode.text),
+        timeoutDataCode: durationTryParse(timeoutDataCode.text),
+        waitUdpSending: durationTryParse(waitUdpSending.text),
+        waitUdpReceiving: durationTryParse(waitUdpReceiving.text),
+        repeat: int.tryParse(repeat.text),
+        portListening: int.tryParse(portListening.text),
+        portTarget: int.tryParse(portTarget.text),
+        thresholdSucBroadcastCount:
+            int.tryParse(thresholdSucBroadcastCount.text),
+        expectedTaskResults: int.tryParse(expectedTaskResults.text),
+      ),
     );
   }
 
   Widget form(BuildContext context) {
     Color color = Theme.of(context).primaryColor;
     return Form(
-      key: _formKey,
+      key: formKey,
       child: ListView(
         padding: const EdgeInsets.symmetric(horizontal: 16),
         children: <Widget>[
           Center(
-            child: OutlineButton(
-              highlightColor: Colors.transparent,
-              highlightedBorderColor: color,
+            child: OutlinedButton(
               onPressed: fetchingWifiInfo ? null : fetchWifiInfo,
-              child: fetchingWifiInfo
-                  ? Text(
-                      'Fetching WiFi info',
-                      style: TextStyle(color: Colors.grey),
-                    )
-                  : Text(
-                      'Use current Wi-Fi',
-                      style: TextStyle(color: color),
-                    ),
+              child: Text(
+                fetchingWifiInfo ? 'Fetching WiFi info' : 'Use current Wi-Fi',
+              ),
             ),
           ),
           TextFormField(
-            controller: _ssid,
+            controller: ssid,
             decoration: const InputDecoration(
               labelText: 'SSID',
               hintText: 'Tony\'s iPhone',
@@ -194,7 +138,7 @@ class _MyAppState extends State<MyApp> {
             ),
           ),
           TextFormField(
-            controller: _bssid,
+            controller: bssid,
             decoration: const InputDecoration(
               labelText: 'BSSID',
               hintText: '00:a0:c9:14:c8:29',
@@ -202,7 +146,7 @@ class _MyAppState extends State<MyApp> {
             ),
           ),
           TextFormField(
-            controller: _password,
+            controller: password,
             decoration: const InputDecoration(
               labelText: 'Password',
               hintText: r'V3Ry.S4F3-P@$$w0rD',
@@ -212,33 +156,33 @@ class _MyAppState extends State<MyApp> {
           RadioListTile(
             title: Text('Broadcast'),
             value: ESPTouchPacket.broadcast,
-            groupValue: _packet,
+            groupValue: packet,
             onChanged: setPacket,
             activeColor: color,
           ),
           RadioListTile(
             title: Text('Multicast'),
             value: ESPTouchPacket.multicast,
-            groupValue: _packet,
+            groupValue: packet,
             onChanged: setPacket,
             activeColor: color,
           ),
           TaskParameterDetails(
             color: color,
-            expectedTaskResults: _expectedTaskResults,
-            intervalGuideCode: _intervalGuideCode,
-            intervalDataCode: _intervalDataCode,
-            timeoutGuideCode: _timeoutGuideCode,
-            timeoutDataCode: _timeoutDataCode,
-            repeat: _repeat,
-            portListening: _portListening,
-            portTarget: _portTarget,
-            waitUdpReceiving: _waitUdpReceiving,
-            waitUdpSending: _waitUdpSending,
-            thresholdSucBroadcastCount: _thresholdSucBroadcastCount,
+            expectedTaskResults: expectedTaskResults,
+            intervalGuideCode: intervalGuideCode,
+            intervalDataCode: intervalDataCode,
+            timeoutGuideCode: timeoutGuideCode,
+            timeoutDataCode: timeoutDataCode,
+            repeat: repeat,
+            portListening: portListening,
+            portTarget: portTarget,
+            waitUdpReceiving: waitUdpReceiving,
+            waitUdpSending: waitUdpSending,
+            thresholdSucBroadcastCount: thresholdSucBroadcastCount,
           ),
           Center(
-            child: RaisedButton(
+            child: ElevatedButton(
               onPressed: () {
                 Navigator.push(
                   context,
@@ -255,65 +199,66 @@ class _MyAppState extends State<MyApp> {
     );
   }
 
-  void setPacket(ESPTouchPacket packet) {
+  void setPacket(ESPTouchPacket? packet) {
+    if (packet == null) return;
     setState(() {
-      _packet = packet;
+      this.packet = packet;
     });
   }
 }
 
 class TaskRoute extends StatefulWidget {
+  TaskRoute({required this.task});
+
   final ESPTouchTask task;
 
-  TaskRoute({this.task});
-
   @override
-  State<StatefulWidget> createState() {
-    return TaskRouteState();
-  }
+  State<StatefulWidget> createState() => TaskRouteState();
 }
 
 class TaskRouteState extends State<TaskRoute> {
-  Stream<ESPTouchResult> _stream;
-  StreamSubscription<ESPTouchResult> _streamSubscription;
-  Timer _timer;
+  late final Stream<ESPTouchResult> stream;
+  late final StreamSubscription<ESPTouchResult> streamSubscription;
+  late final Timer timer;
 
-  final List<ESPTouchResult> _results = [];
+  final List<ESPTouchResult> results = [];
 
   @override
   void initState() {
-    _stream = widget.task.execute();
-    _streamSubscription = _stream.listen(_results.add);
+    stream = widget.task.execute();
+    streamSubscription = stream.listen(results.add);
     final receiving = widget.task.taskParameter.waitUdpReceiving;
     final sending = widget.task.taskParameter.waitUdpSending;
     final cancelLatestAfter = receiving + sending;
-    _timer = Timer(cancelLatestAfter, () {
-      _streamSubscription?.cancel();
-      if (_results.isEmpty && mounted) {
-        showDialog(
+    timer = Timer(
+      cancelLatestAfter,
+      () {
+        streamSubscription.cancel();
+        if (results.isEmpty && mounted) {
+          showDialog(
             context: context,
             builder: (context) {
               return AlertDialog(
                 title: Text('No devices found'),
                 actions: <Widget>[
-                  FlatButton(
-                    onPressed: () {
-                      Navigator.of(context)..pop()..pop();
-                    },
+                  TextButton(
+                    onPressed: () => Navigator.of(context)..pop()..pop(),
                     child: Text('OK'),
                   ),
                 ],
               );
-            });
-      }
-    });
+            },
+          );
+        }
+      },
+    );
     super.initState();
   }
 
   @override
   dispose() {
-    _timer.cancel();
-    _streamSubscription?.cancel();
+    timer.cancel();
+    streamSubscription.cancel();
     super.dispose();
   }
 
@@ -340,8 +285,8 @@ class TaskRouteState extends State<TaskRoute> {
   copyValue(BuildContext context, String label, String v) {
     return () {
       Clipboard.setData(ClipboardData(text: v));
-      Scaffold.of(context).showSnackBar(
-          SnackBar(content: Text('Copied $label to clipboard: $v')));
+      final snackBar = SnackBar(content: Text('Copied $label to clipboard $v'));
+      ScaffoldMessenger.of(context).showSnackBar(snackBar);
     };
   }
 
@@ -351,9 +296,10 @@ class TaskRouteState extends State<TaskRoute> {
 
   Widget resultList(BuildContext context) {
     return ListView.builder(
-      itemCount: _results.length,
+      itemCount: results.length,
       itemBuilder: (_, index) {
-        final result = _results[index];
+        final result = results[index];
+        final textTheme = Theme.of(context).textTheme;
         return Padding(
           padding: const EdgeInsets.all(8.0),
           child: Row(
@@ -363,7 +309,7 @@ class TaskRouteState extends State<TaskRoute> {
                 onLongPress: copyValue(context, 'BSSID', result.bssid),
                 child: Row(
                   children: <Widget>[
-                    Text('BSSID: ', style: Theme.of(context).textTheme.body2),
+                    Text('BSSID: ', style: textTheme.bodyText1),
                     Text(result.bssid,
                         style: TextStyle(fontFamily: 'monospace')),
                   ],
@@ -373,7 +319,7 @@ class TaskRouteState extends State<TaskRoute> {
                 onLongPress: copyValue(context, 'IP', result.ip),
                 child: Row(
                   children: <Widget>[
-                    Text('IP: ', style: Theme.of(context).textTheme.body2),
+                    Text('IP: ', style: textTheme.bodyText1),
                     Text(result.ip, style: TextStyle(fontFamily: 'monospace')),
                   ],
                 ),
@@ -398,10 +344,10 @@ class TaskRouteState extends State<TaskRoute> {
               return error(context, 'Error in StreamBuilder');
             }
             if (!snapshot.hasData) {
+              final primaryColor = Theme.of(context).primaryColor;
               return Center(
                 child: CircularProgressIndicator(
-                  valueColor:
-                      AlwaysStoppedAnimation(Theme.of(context).primaryColor),
+                  valueColor: AlwaysStoppedAnimation(primaryColor),
                 ),
               );
             }
@@ -415,11 +361,182 @@ class TaskRouteState extends State<TaskRoute> {
               case ConnectionState.waiting:
                 return waitingState(context);
             }
-            return error(context, 'Unexpected');
           },
-          stream: _stream,
+          stream: stream,
         ),
       ),
     );
   }
+}
+
+class TaskParameterDetails extends StatelessWidget {
+  const TaskParameterDetails({
+    Key? key,
+    required this.color,
+    required this.expectedTaskResults,
+    required this.intervalGuideCode,
+    required this.intervalDataCode,
+    required this.timeoutGuideCode,
+    required this.timeoutDataCode,
+    required this.repeat,
+    required this.portListening,
+    required this.portTarget,
+    required this.waitUdpReceiving,
+    required this.waitUdpSending,
+    required this.thresholdSucBroadcastCount,
+  }) : super(key: key);
+
+  final Color color;
+  final TextEditingController expectedTaskResults;
+  final TextEditingController intervalGuideCode;
+  final TextEditingController intervalDataCode;
+  final TextEditingController timeoutGuideCode;
+  final TextEditingController timeoutDataCode;
+  final TextEditingController repeat;
+  final TextEditingController portListening;
+  final TextEditingController portTarget;
+  final TextEditingController waitUdpReceiving;
+  final TextEditingController waitUdpSending;
+  final TextEditingController thresholdSucBroadcastCount;
+
+  @override
+  Widget build(BuildContext context) {
+    return ExpansionTile(
+      title: Text(
+        'Task parameter detail',
+        style: TextStyle(color: color),
+      ),
+      trailing: Icon(
+        Icons.settings,
+        color: color,
+      ),
+      initiallyExpanded: false,
+      children: <Widget>[
+        OptionalIntegerTextField(
+          controller: expectedTaskResults,
+          labelText: 'Expected task results (count)',
+          hintText: 1,
+          helperText: 'The number of devices you expect to scan.',
+        ),
+        OptionalIntegerTextField(
+          controller: intervalGuideCode,
+          labelText: 'Interval guide code (ms)',
+          hintText: 8,
+        ),
+        OptionalIntegerTextField(
+          controller: intervalDataCode,
+          labelText: 'Interval data code (ms)',
+          hintText: 8,
+        ),
+        OptionalIntegerTextField(
+          controller: timeoutGuideCode,
+          labelText: 'Timeout guide code (ms)',
+          hintText: 2000,
+        ),
+        OptionalIntegerTextField(
+          controller: timeoutDataCode,
+          labelText: 'Timeout data code (ms)',
+          hintText: 4000,
+        ),
+        OptionalIntegerTextField(
+          controller: repeat,
+          labelText: 'Repeat',
+          hintText: 1,
+        ),
+        // The mac and ip length are skipped for now.
+        OptionalIntegerTextField(
+          controller: portListening,
+          labelText: 'Listen on port',
+          hintText: 18266,
+        ),
+        OptionalIntegerTextField(
+          controller: portTarget,
+          labelText: 'Target port',
+          hintText: 7001,
+        ),
+        OptionalIntegerTextField(
+          controller: waitUdpReceiving,
+          labelText: 'Wait UDP receiving (ms)',
+          hintText: 15000,
+        ),
+        OptionalIntegerTextField(
+          controller: waitUdpSending,
+          labelText: 'Wait UDP sending (ms)',
+          hintText: 45000,
+        ),
+        OptionalIntegerTextField(
+          controller: thresholdSucBroadcastCount,
+          labelText: 'Broadcast count success threshold',
+          hintText: 1,
+        ),
+      ],
+    );
+  }
+}
+
+class OptionalIntegerTextField extends StatelessWidget {
+  const OptionalIntegerTextField({
+    Key? key,
+    required this.controller,
+    required this.labelText,
+    required this.hintText,
+    this.helperText,
+  }) : super(key: key);
+
+  final TextEditingController controller;
+  final String labelText;
+  final int hintText;
+  final String? helperText;
+
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      controller: controller,
+      keyboardType: TextInputType.number,
+      validator: (String? value) {
+        // The user didn't edit the field, so can return early without an error
+        // and skip validating (int parsing).
+        value = (value ?? '').trim();
+        if (value == '') return null;
+
+        final v = int.tryParse(value, radix: 10);
+        if (v == null) return 'Please enter an integer number';
+      },
+      autovalidateMode: AutovalidateMode.onUserInteraction,
+      decoration: InputDecoration(
+        labelText: labelText,
+        hintText: hintText.toString(),
+        helperText: helperText,
+      ),
+    );
+  }
+}
+
+const helperSSID = "SSID is the technical term for a network name. "
+    "When you set up a wireless home network, "
+    "you give it a name to distinguish it from other networks in your neighbourhood.";
+const helperBSSID =
+    "BSSID is the MAC address of the wireless access point (router).";
+const helperPassword = "The password of the Wi-Fi network";
+
+/// The method channel is used to get the SSID and BSSID.
+///
+/// For a real app, consider using the
+/// [`connectivity`](https://pub.dev/packages/connectivity) and
+/// [`wifi_info_flutter`](https://pub.dev/packages/wifi_info_flutter) packages.
+///
+/// For more info: https://github.com/smaho-engineering/esptouch_flutter/blob/master/example/README.md#get-wifi-details
+class SimpleWifiInfo {
+  static const platform =
+      MethodChannel('eng.smaho.com/esptouch_plugin/example');
+
+  /// Get WiFi SSID using platform channels.
+  ///
+  /// Can return null if BSSID information is not available.
+  static Future<String?> get ssid => platform.invokeMethod('ssid');
+
+  /// Get WiFi BSSID using platform channels.
+  ///
+  /// Can return null if BSSID information is not available.
+  static Future<String?> get bssid => platform.invokeMethod('bssid');
 }